refactor: simplify match statements and improve code clarity
- Streamlined match statements in `run_cache_subscription` for better readability.
- Adjusted account instruction handling in `geyser_listener` to maintain type consistency.
- Removed unnecessary default initialization in `tpu_client` configuration, enhancing code efficiency.

Files affected: `src/execution/cache_geyser.rs`, `src/solana/geyser_listener.rs`, `src/solana/tpu_client.rs`

+++ src/solana/geyser_listener.rs
@@ -524,7 +524,7 @@ impl GeyserListener {
                                             for inner_ix in &inner_ix_group.instructions {
                                                 inner_instructions.push(InnerInstruction {
                                                     program_id_index: inner_ix.program_id_index as u8,
-                                                    accounts: inner_ix.accounts.iter().map(|&a| a as u8).collect(),
+                                                    accounts: inner_ix.accounts.iter().map(|&a| a).collect(),
                                                     data: inner_ix.data.clone(),
                                                 });
                                             }
